requirements.txt
attrs==19.3.0
aws-cdk.assets==1.33.0
aws-cdk.aws-cloudformation==1.32.1
aws-cdk.aws-cloudwatch==1.33.0
aws-cdk.aws-directoryservice==1.32.1
aws-cdk.aws-ec2==1.33.0
aws-cdk.aws-events==1.33.0
aws-cdk.aws-iam==1.33.0
aws-cdk.aws-kms==1.33.0
aws-cdk.aws-lambda==1.33.0
aws-cdk.aws-logs==1.33.0
aws-cdk.aws-route53==1.32.1
aws-cdk.aws-s3==1.33.0
aws-cdk.aws-s3-assets==1.33.0
aws-cdk.aws-sam==1.33.0
aws-cdk.aws-secretsmanager==1.33.0
aws-cdk.aws-sns==1.32.1
aws-cdk.aws-sqs==1.33.0
aws-cdk.aws-ssm==1.33.0
aws-cdk.aws-workspaces==1.32.1
aws-cdk.cloud-assembly-schema==1.33.0
aws-cdk.core==1.33.0
aws-cdk.cx-api==1.33.0
aws-cdk.region-info==1.33.0
boto3==1.10.5
botocore==1.13.5
cattrs==1.0.0
constructs==3.0.1
docutils==0.15.2
jmespath==0.9.4
jsii==1.3.1
publication==0.0.3
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
typing-extensions==3.7.4.1
urllib3==1.25.9
